Google Cloud Functions/requirements.txt
pandas==0.25.1
requests==2.22.0
pyarrow==0.14.1
google-cloud-bigquery==1.18.0
